Little changes MesaExamen

+++ resources/views/MesaExamen/index.blade.php
@@ -42,14 +42,23 @@
                             </th>
                             <td>
                                 <div class="btn-group">
-                                    <a href="/materias/{{ $mesa->id }}/edit">
-                                        <button type="button" class="btn btn-primary btn-small">Editar</button>
-                                    </a>
-                                    <button type="button" class="btn red lighten-1 ml-2 btn-small"
-                                            data-toggle="modal" data-target="#deleteModal"
-                                            onclick="deleteSignature($(this))"
-                                            data-name="{{$mesa->name}}" data-id="{{$mesa->id}}">Borrar
-                                    </button>
+                                    @if(auth()->user()->isAdmin())
+                                        <a href="/materias/{{ $mesa->id }}/edit">
+                                            <button type="button" class="btn btn-primary btn-small">Editar</button>
+                                        </a>
+                                        <button type="button" class="btn red lighten-1 ml-2 btn-small"
+                                                data-toggle="modal" data-target="#deleteModal"
+                                                onclick="deleteSignature($(this))"
+                                                data-name="{{$mesa->name}}" data-id="{{$mesa->id}}">Borrar
+                                        </button>
+                                    @endif
+                                    @if(auth()->user()->isStudent())
+                                        <a href="/materias/{{ $mesa->id }}/edit">
+                                            <button type="button" class="btn btn-primary btn-small">Editar</button>
+                                        </a>
+
+                                    @endif
+                                   
                                 </div>
                             </td>
                         </tr>
